CPU Debug modified to show also last address bus and data bus values. The way the latched configuration of the ROM in C264 series is managed is also changed. Improving C264 series.

+++ src/C264/TED.cpp
@@ -33,7 +33,7 @@ bool C264::TED::simulate (MCHEmul::CPU* cpu)
 
 	if (_TEDRegisters -> romAccessChanged ()) // Changes the value...
 		notify (MCHEmul::Event (_ROMACCESSCHANGED, 
-			(unsigned int) _TEDRegisters -> hiromSelected () ? 1 : 0));
+			_TEDRegisters -> hiromSelected () ? 1 : 0));
 
 	return (true);
 }
